Add user authentication for inventory views
Add login_required methods and update querysets to implement user authentication for all views. Users that are logged in only have access those products and related information that are associated with their account. Users that aren't logged in can't see any products, and will be prompted to log in or create a new account.