Fix NoReverseMatch error
Fix error that throws a NoReverseMatch exception when trying to resolve the Product detail url after updating that product with the new ProductUpdate view.

The error was caused by an improper keyword 'label' being passed to the urlconf by the model's get_absolute_url method, instead of 'product' as it is in the path for the product_detail view.